changed input and output, added fee handling (not tested)

+++ main.hpp
@@ -120,18 +120,12 @@
 #define TXS_BNK_FEE      (0x10000000) /* 0x0.1 G */
 #define TXS_BKY_FEE      (0x10000000) /* 0x0.1 G */
 #define USER_MIN_MASS    (0x10000000) /* 0x0.1 G minimum user account mass to send transaction */
-//#define BANK_MIN_MASS    (0x100000000)
-//#define BANK_MIN_WEIGHT  (0x1000000000)
 #define BANK_MIN_UMASS   (0x10000000) /* 0x0.1G, minimum admin account mass to send transaction */
 #define BANK_MIN_TMASS   (0x1000000000) /* 0x10G, if bank total mass below this value, bank can be taken over */
 #define BANK_MIN_MTIME   (0x200*BLOCKSEC) /* if no transaction in this period bank can be taken over */
 #define BANK_MAX         (0xffff)
 #define BANK_PROFIT(x)   ((x)>>4) /* 1/16 of fees */
-#define BANK_USER_FEE(x) (0x1000 + ((BANK_PROFIT((x)*TXS_DIV_FEE/BLOCKDIV))>>2)) /* collected every block */
-//#define BANK_TIME_FEE(x) (0x40*(x)) /* must have minimum of 512 accounts to make profits */
-//#define START_AGE      (0x100*BLOCKSEC) /* 100 blocks fee */
-//#define TIME_FEE(x,y) (0x10*((x)-(y))/BLOCKSEC) /* seconds */
-//#define TXS_STP_FEE    (0x1000) /* minimum 1hour fee */
+#define BANK_USER_FEE(x) (0x1000 + ((BANK_PROFIT(((uint64_t)(x))*(TXS_DIV_FEE/BLOCKDIV)))>>2)) /* every block */
 
 #define MESSAGE_TOO_LONG 0x800000
 #define MESSAGE_LEN_OK 0x10000
